apt-spider/requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.1
astor==0.8.1
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
blinker==1.4
bpemb==0.3.3
cachetools==5.0.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.4
conllu==4.4.1
cryptography==3.4.7
cssselect==1.1.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.1.7
django-utils-six==2.0
djangorestframework==3.12.4
emoji==1.6.3
entrypoints==0.4
executing==0.8.2
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
flair==0.10
fonttools==4.29.1
ftfy==6.1.1
gdown==3.12.2
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.2
gunicorn==19.7.1
h11==0.12.0
h2==4.0.0
hpack==4.0.0
huggingface-hub==0.4.0
hyperframe==6.0.1
idna==2.10
IMDbPY==2021.4.18
importlib-metadata==3.10.1
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
Janome==0.4.2
jedi==0.18.1
jieba3k==0.35.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaitaistruct==0.9
kiwisolver==1.3.2
konoha==4.6.5
langdetect==1.0.9
lxml==4.8.0
MarkupSafe==2.1.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
more-itertools==8.8.0
mpld3==0.3
mypy-extensions==0.4.3
mysql-connector-python==8.0.25
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
newspaper3k==0.2.8
nltk==3.7
notebook==6.4.8
numpy==1.22.2
overrides==3.1.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
plotly==5.6.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.17.3
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pycparser==2.20
pydeck==0.7.1
pydivert==2.1.0
Pygments==2.11.2
Pympler==1.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==22.3.0
random-user-agent==1.0.1
regex==2022.1.18
requests==2.25.1
requests-file==1.5.1
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.8.0
segtok==1.5.11
selenium==3.141.0
selenium-wire==4.3.1
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.95
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.1
SQLAlchemy==1.4.39
sqlitedict==2.0.0
sqlparse==0.4.1
stack-data==0.2.0
streamlit==1.6.0
tabulate==0.8.9
tenacity==8.0.1
terminado==0.13.1
testpath==0.6.0
text2emotion==0.0.5
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.2.0
tokenizers==0.11.6
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.10.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.17.0
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.4
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==3.3.1
widgetsnbextension==3.5.2
Wikipedia-API==0.5.4
wrapt==1.13.3
wsproto==1.0.0
zipp==3.7.0